// frontend/src/app/features/perfiles/models/profile-status.ts
export enum ProfileStatus {
  SolicitudAprobada = 'SolicitudAprobada',
  PendienteGeneracionPerfil = 'PendienteGeneracionPerfil',
  EnRevisionRRHHPE = 'EnRevisionRRHHPE',
  ResumenEjecutivoGenerado = 'ResumenEjecutivoGenerado',
  EnRevisionAreaSol = 'EnRevisionAreaSol',
  Observada = 'Observada',
  Corregida = 'Corregida',
  Aprobada = 'Aprobada',
  PerfilAprobadoFinal = 'PerfilAprobadoFinal',
}